Start splitting convert_osm into a piece that only produces a StreetNetwork.
Only partly done. The crux of this refactor, extracting stuff from OSM, remains...
