+++ convert_osm/src/lib.rs
@@ -6,94 +6,23 @@ extern crate log;
 use std::collections::{HashMap, HashSet};
 
 use anyhow::Result;
-use serde::{Deserialize, Serialize};
 
 use abstio::MapName;
 use abstutil::{Tags, Timer};
 use geom::{Distance, FindClosest, GPSBounds, HashablePt2D, LonLat, PolyLine, Polygon, Pt2D, Ring};
-use raw_map::{osm, Amenity, MapConfig, OriginalRoad, RawMap, RawRoad};
+use raw_map::{osm, Amenity, OriginalRoad, RawMap, RawRoad};
+
+pub use import_streets::{
+    OnstreetParking, Options, PrivateOffstreetParking, PublicOffstreetParking,
+};
 
 mod clip;
 mod elevation;
 mod extract;
 mod gtfs;
-pub mod osm_geom;
+mod import_streets;
+mod osm_reader;
 mod parking;
-pub mod reader;
-mod split_ways;
-
-/// Configures the creation of a RawMap from OSM and other input data.
-pub struct Options {
-    pub map_config: MapConfig,
-
-    pub onstreet_parking: OnstreetParking,
-    pub public_offstreet_parking: PublicOffstreetParking,
-    pub private_offstreet_parking: PrivateOffstreetParking,
-    /// OSM railway=rail will be included as light rail if so. Cosmetic only.
-    pub include_railroads: bool,
-    /// If provided, read polygons from this GeoJSON file and add them to the RawMap as buildings.
-    pub extra_buildings: Option<String>,
-    /// Only include highways and arterials. This may make sense for some region-wide maps for
-    /// particular use cases.
-    pub skip_local_roads: bool,
-    /// Only include crosswalks that match a `highway=crossing` OSM node.
-    pub filter_crosswalks: bool,
-    /// Configure public transit using this URL to a static GTFS feed in .zip format.
-    pub gtfs_url: Option<String>,
-    pub elevation: bool,
-}
-
-impl Options {
-    pub fn default_for_side(driving_side: raw_map::DrivingSide) -> Self {
-        Self {
-            map_config: MapConfig::default_for_side(driving_side),
-            onstreet_parking: OnstreetParking::JustOSM,
-            public_offstreet_parking: PublicOffstreetParking::None,
-            private_offstreet_parking: PrivateOffstreetParking::FixedPerBldg(1),
-            include_railroads: true,
-            extra_buildings: None,
-            skip_local_roads: false,
-            filter_crosswalks: false,
-            gtfs_url: None,
-            elevation: false,
-        }
-    }
-}
-
-/// What roads will have on-street parking lanes? Data from
-/// <https://wiki.openstreetmap.org/wiki/Key:parking:lane> is always used if available.
-#[derive(Clone, Serialize, Deserialize)]
-pub enum OnstreetParking {
-    /// If not tagged, there won't be parking.
-    JustOSM,
-    /// If OSM data is missing, then try to match data from
-    /// <http://data-seattlecitygis.opendata.arcgis.com/datasets/blockface>. This is Seattle specific.
-    Blockface(String),
-    /// If OSM data is missing, then infer parking lanes on some percentage of
-    /// "highway=residential" roads.
-    SomeAdditionalWhereNoData {
-        /// [0, 100]
-        pct: usize,
-    },
-}
-
-/// How many spots are available in public parking garages?
-#[derive(Clone, Serialize, Deserialize)]
-pub enum PublicOffstreetParking {
-    None,
-    /// Pull data from
-    /// <https://data-seattlecitygis.opendata.arcgis.com/datasets/public-garages-or-parking-lots>, a
-    /// Seattle-specific data source.
-    Gis(String),
-}
-
-/// If a building doesn't have anything from public_offstreet_parking and isn't tagged as a garage
-/// in OSM, how many private spots should it have?
-#[derive(Clone, Serialize, Deserialize)]
-pub enum PrivateOffstreetParking {
-    FixedPerBldg(usize),
-    // TODO Based on the number of residents?
-}
 
 /// Create a RawMap from OSM and other input data.
 pub fn convert(
@@ -114,15 +43,16 @@ pub fn convert(
         map.streets.gps_bounds = gps_bounds;
     }
 
-    let extract = extract::extract_osm(&mut map, &osm_input_path, clip_path, &opts, timer);
-    let split_output = split_ways::split_up_roads(&mut map, extract, timer);
+    let (extract, amenity_points) =
+        extract::extract_osm(&mut map, &osm_input_path, clip_path, &opts, timer);
+    let split_output = import_streets::split_ways::split_up_roads(&mut map.streets, extract, timer);
     clip::clip_map(&mut map, timer);
 
     // Need to do a first pass of removing cul-de-sacs here, or we wind up with loop PolyLines when
     // doing the parking hint matching.
     map.streets.roads.retain(|r, _| r.i1 != r.i2);
 
-    use_amenities(&mut map, split_output.amenities, timer);
+    use_amenities(&mut map, amenity_points, timer);
 
     parking::apply_parking(&mut map, &opts, timer);
