+++ convert_osm/src/osm_reader/reader.rs
@@ -5,7 +5,7 @@ use anyhow::Result;
 use abstio::slurp_file;
 use abstutil::{prettyprint_usize, Tags, Timer};
 use geom::{GPSBounds, LonLat, Pt2D};
-use raw_map::osm::{NodeID, OsmID, RelationID, WayID};
+use street_network::osm::{NodeID, OsmID, RelationID, WayID};
 
 // References to missing objects are just filtered out.
 // Per https://wiki.openstreetmap.org/wiki/OSM_XML#Certainties_and_Uncertainties, we assume
