+++ convert_osm/src/parking.rs
@@ -1,9 +1,12 @@
 use abstutil::Timer;
 use geom::{Distance, FindClosest, PolyLine};
 use kml::ExtraShapes;
-use raw_map::{osm, OriginalRoad, RawMap};
+use raw_map::RawMap;
+use street_network::{osm, OriginalRoad};
 
-use crate::{OnstreetParking, Options, PrivateOffstreetParking, PublicOffstreetParking};
+use crate::import_streets::{
+    OnstreetParking, Options, PrivateOffstreetParking, PublicOffstreetParking,
+};
 
 // Just used for matching hints to different sides of a road.
 const DIRECTED_ROAD_THICKNESS: Distance = Distance::const_meters(2.5);
